Add setOpenInNewTask() to AppSettingsDialog.Builder
When using setOpenInNewTask() for building the dialog, the app settings
screen will be opened using an Intent that has the flag
Intent.FLAG_ACTIVITY_NEW_TASK added before calling
startActivityForResult(). Hence the settings Activity will not appear as
a screen within the task of the activity it has been opened from but
instead as a separate task that can be brought to front and back
separately by the user.